Move hooks to the right folder
