+++ packages/frontend/hooks/scaffold-eth/useAutoConnect.ts
