+++ packages/frontend/hooks/scaffold-eth/useBurnerWallet.ts
